Allow cfitsio to be compiled from source.
If the static feature is supplied, then the cfitsio source code is
compiled at build time.

The source code is taken from the fitsio website (version 3.49). I
removed the docs directory from the output to keep the data volume down.

I've tested this on Linux and it seems to work fine. I don't have access
to OSX to know how it would go over there, but hopefully it also just
works.